Redaction Toolkit | Support to redact some Latin ligature letters and letters with diacritics (#1012)
## Purpose

We currently cannot handle accented letter like é and ligature letter like œ, which is common in French. Need to support this. We also expand the language coverage out of just French. Please check the code for the actual charset we have taken care of.

## Validation

Before merging this PR, please make sure below works are done and marked items with 'x'.

- [x] Your code builds clean without any errors or warnings.
- [x] You have tested your change manually.
- [x] You have added unit tests.

Co-authored-by: Chia-Sheng Chen <chiache@microsoft.com>